move tracker splitting to SupportedData
